refactor(semantictokens): детект пересечений вынесен из провайдера в тест-helper
findOverlaps и record TokenOverlap были detection-only кодом (прод-вызовов нет,
только тесты) в продакшен-классе SemanticTokensProvider — лишняя публичная
поверхность. Перенесены в тестовый TokenOverlaps в том же пакете; провайдер
больше не несёт test-only API. Поведение и тесты не изменились.
